refactor(formatter): extract RequestMessageComposer component
- Create FormattingComponents::RequestMessageComposer for message enrichment
- Extract build_controller_action_message and append_request_uri logic
- Rename append_request_details to include_component_details
- Update JsonFormatter to delegate to new component
- Add require for new formatting_components module
